added support for two types of virtual environments in updater

+++ updater.py
@@ -19,7 +19,7 @@ for root, dirs, files in os.walk(os.getcwd(), topdown=False):
         if name != "updater.py" and (name not in ['analyticsData.txt', 'certificates.txt', 'accessIdentities.txt', 'validOTPCodes.txt', '.env', 'authorisation.txt']) and (not name.startswith('aa-report')) and ("AccessFolders" not in root):
             os.remove(os.path.join(root, name))
     for name in dirs:
-        if name != 'analyticsReports' and name != 'AccessFolders' and ("AccessFolders" not in root):
+        if name != 'analyticsReports' and name != 'AccessFolders' and ("AccessFolders" not in root) and (name not in ['venv', 'virt']):
             os.rmdir(os.path.join(root, name))
 
 time.sleep(2)
